improvements and refactor
