+++ package.json
@@ -19,6 +19,7 @@
     "phosphor-react": "^1.4.1",
     "react": "^18.2.0",
     "react-axios": "^2.0.6",
+    "react-carbonui": "^1.3.1",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.42.1",
     "react-router-dom": "^6.9.0",
