+++ src/components/views/about/About.jsx
@@ -1,6 +1,4 @@
-import React from 'react';
-
-const AboutPage = () => {
+const About = () => {
   return (
     <div>
       <h1>About Us</h1>
@@ -10,4 +8,4 @@ const AboutPage = () => {
   );
 };
 
-export default AboutPage;
+export default About;
