+++ src/components/views/contact/Contact.jsx
@@ -1,0 +1,13 @@
+import { Button } from 'react-carbonui';
+
+const Contact = () => {
+  return (
+    <div>
+      <h1>Contact Us</h1>
+
+      <Button size="large">Coming soon</Button>
+    </div>
+  );
+};
+
+export default Contact;
