--- src/pages/ContactPage.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const ContactPage = () => {
-  return (
-    <div>
-      <h1>Contact Us</h1>
-      <h2> Coming soon!!</h2>
-    </div>
-  );
-};
-
-export default ContactPage;
