--- src/pages/Root.jsx
@@ -1,11 +1,0 @@
-const Root = () => {
-  return (
-    <main>
-      Hi asahs
-      {/* <Outlet /> */}
-      <button className="ct_langmode_btn">asas </button>
-    </main>
-  );
-};
-
-export default Root;
